move helper function to Util
